ai-order-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests

langchain
langchain-community
langchain-openai
chromadb

openai
pandas
aiofiles
aiosqlite
